refactor: Correct StrenApp's package location
